Item upload fix, shop creation fix, item upload ui layout tweak

+++ app/src/main/java/com/example/sagar/popupshops_buyerside/Shop/Item.java
@@ -3,7 +3,7 @@ package com.example.sagar.popupshops_buyerside.Shop;
 public class Item {
 
     private String itemCategory;
-    private int itemPrice;
+    private float itemPrice;
     private String itemDescription;
     private String itemImage;
     private int itemStock;
@@ -12,7 +12,7 @@ public class Item {
 
     }
 
-    public Item(String itemCategory, int itemPrice, String itemDescription, String itemImage, int itemStock) {
+    public Item(String itemCategory, float itemPrice, String itemDescription, String itemImage, int itemStock) {
         this.itemCategory = itemCategory;
         this.itemPrice = itemPrice;
         this.itemDescription = itemDescription;
@@ -24,7 +24,7 @@ public class Item {
         return 6;
     }
 
-    public int getItemPrice() {
+    public float getItemPrice() {
         return itemPrice;
     }
